finish

+++ wgsMLVA.py
@@ -13,23 +13,29 @@ from Bio import SeqIO
 #load in the VNTR chart data.
 
 parser = argparse.ArgumentParser(prog = " MLVA in silico Program", \
-								usage = 'Find MLVA code from fasta sequence.')
-parser.add_argument('-i', '--inputFile', required = True , \
+								usage = 'Find MLVA code from closed \
+										fasta sequence.')
+parser.add_argument('-in', '--inputFile', required = True , \
 					help = 'input fasta file. Required.')
-parser.add_argument('-o', '--outputDir', required = True, \
+parser.add_argument('-out', '--outputDir', required = True, \
 					help = 'MLVA output directory. Required.')
 parser.add_argument('-v', '--version', action = 'version', \
-					version = '%(prog)s v0.1')
+					version = '%(prog)s version 1.0')
 parser.add_argument('-n', '--name', required = False, \
 					help = 'Assign file name. If not assigned it will use \
 							the file name of the fasta file.')
-parser.add_argument('-u', '--unmask', action = "store_true", default= False)
+parser.add_argument('-u', '--unmask', action = "store_true", default= False, \
+					help = 'Removes VNTR3 masking, gives actual VNTR3 \
+							repeat values.')
+parser.add_argument('-k', '--keepTemp', action = "store_true", default= False, \
+					help = 'Keeps single-line temporary fasta file.')
 args = parser.parse_args()
 inFile = args.inputFile
 outDir = args.outputDir
 mlvaDir = os.path.dirname(os.path.realpath(sys.argv[0]))
 sampleID = args.name
 fix = args.unmask
+keep = args.keepTemp
 
 #------------------------------------------------------------------------------
 #Check parsed arguments
@@ -42,23 +48,13 @@ else:
 	color_red = '\33[31m'
 	print(color_red + "\n>>> Error: Please input fasta\n" + color_red)
 	sys.exit()
-####------------------------------------------------------------------------------
-#Various optional parse_args (WIP)
-
-def quiet():
-	d = open("%sMLVA/%s" % (outDir, MLVAfile), 'r')
-	print d
-####------------------------------------------------------------------------------
+#------------------------------------------------------------------------------
 #Stores variables for program.
 #Variable "name" changes the name of the files and directories created.
 if sampleID:
 	name = sampleID
 else:	
-	#pullName = inFile.rsplit('/', 1)[1]
-	print inFile
-	prepName = re.findall("(\w\d\d\d)",inFile)
-	name = prepName[0]
-	#pullName.split('.', 1)[0]
+	name = re.findall("(\w\d\d\d)",inFile)[0]
 	outFile = name + "fasta_out"
 	outSummary = name + "summary"
 #------------------------------------------------------------------------------
@@ -68,14 +64,13 @@ if not os.path.exists("%s/MLVA" % outDir):
 	os.mkdir("%s/MLVA" % outDir)
 if not os.path.exists("%s/MLVA/summary.txt" % outDir):
 	with open("%s/MLVA/summary.txt" % outDir, "a") as summary:
-		summary.write("Isolate,MLVA_number[VNTR1, \
-			VNTR3_1, VNTR3_2, VNTR4, VNTR5, VNTR6]\n")
+			summary.write("Isolate: MLVA_number [VNTR1, VNTR3a, VNTR3b, VNTR4, VNTR5, VNTR6]\n")	
 if not os.path.exists("%s/MLVA/MLVA_data" % outDir):
 	os.mkdir("%s/MLVA/MLVA_data" % outDir)		
 if not os.path.exists("%s/MLVA/sequence_data" % outDir):
 	os.mkdir("%s/MLVA/sequence_data" % outDir)
 #------------------------------------------------------------------------------
-print "\n>>> Opening FASTA file..."
+print "\n>>> Opening", name, "fasta file..."
 
 #Reads sequence file list and stores it as a string object. Close file.
 #Splits string at the start of a line.
@@ -86,7 +81,6 @@ with open(inFile,"r") as newFile:
 	sequences = re.split("^>", sequences, flags = re.MULTILINE)
 	del sequences[0]
 	newFile.close()
-print ">>> Converting FASTA file..."
 #Conversts multiline fasta to single line. Writes new fasta to file: "X_out".
 #1. Split each fasta into header and sequence.
 #2. Replace ">" lost in ">" split, Replace "\n" lost in split directly above.
@@ -101,7 +95,6 @@ with open("%s/MLVA/MLVA_data/%s" % (outDir, outFile),"w") as newFasta:
 		sequence = sequence.replace("\n","") + "\n"
 		newFasta.write(sequence)
 	newFasta.close()
-print ">>> Multiline FASTA is now a Single Line FASTA."
 #------------------------------------------------------------------------------
 #Each primer and reverse primer for genome search. 
 #V(1 to 6)p(1 or 2)(start or end) = VNTR(1 to 6)primer(1 or 2)(start or end)
@@ -125,6 +118,7 @@ V6p1s = "CCAACGGCGGTCTGCTGGGTGGTC"
 V6p1e = "GGTAGCGGCGCAGCGGGCGGCG"
 V6p2s = "CGCCGCCCGCTGCGCCGCTACC"
 V6p2e = "GACCACCCAGCAGACCGCCGTTGG"
+
 #VNTR repeats
 VNTR1wt1 = "CTGCTTGGCGGGTTC"
 VNTR1wt2 = "GAACCCGCCAAGCAG"
@@ -141,25 +135,20 @@ VNTR5_1 = "TGGTGC"
 VNTR5_2 = "GCACCA"
 VNTR6_1 = "GGCGGCTCG"
 VNTR6_2 = "CGAGCCGCC"
-#List of MLVA repeat numbers.
+
 #Order: [VNTR1, VNTR3_1, VNTR3b_2, VNTR4, VNTR5, VNTR6]
 MLVA_code = []
 
 #Creating and opening output file.
-
-MLVAfile = name + "_MLVA.txt"
+MLVAfile = name + "MLVA.txt"
 q = open("%s/MLVA/MLVA_data/%s" % (outDir, MLVAfile), "w")
 seq = open("%s/MLVA/sequence_data/%s" % (outDir, name), "a")
 
-
-#t = open("%s/MLVA/MLVA_numbers.txt" % outDir, "a")
 print >>q, name
-
 with open("%s/MLVA/MLVA_data/%s" % (outDir, outFile), 'r') as fasta:
 	all_dna = fasta.read()
 
-
-print "\n>>> Starting MLVA Search..."
+print ">>> Starting MLVA Search..."
 
 #------------------------------------------------------------------------------
 #attempt at doing a fuzzy repeat search. Still WIP.
@@ -172,9 +161,9 @@ def fuzzy_search(search_key, text, strictness):
 
 #------------------------------------------------------------------------------
 #all VNTR search scripts. 
-def VNTR1_search(Name, p1s, p1e, p2s, p2e, r1, r2, r3, r4, r5, r6):
+def VNTR1_search(head, p1s, p1e, p2s, p2e, r1, r2, r3, r4, r5, r6):
 
-	print>>q, "\n >>>" + Name + "\n"
+	print>>q, "\n>>> " + head
 
 	if p1s and p1e in all_dna:
 		for x in re.finditer(p1s, all_dna):
@@ -183,9 +172,6 @@ def VNTR1_search(Name, p1s, p1e, p2s, p2e, r1, r2, r3, r4, r5, r6):
 			fp2 = int(y.end())
 		fDistance = fp2 - fp1
 		fragment = all_dna[fp1:fp2]
-		print >>seq, ">%s" % (Name), "\n" + fragment
-		print >>q, "VNTR Fragment locus:", fp1, "-", fp2
-		print >>q, "VNTR fragment size:", fDistance
 		if r1 in fragment:
 			#print fuzzy_search(r1, fragment, 0.85)
 			#vhtr_count = fuzzy_search(r1, fragment, 0.85)
@@ -199,6 +185,10 @@ def VNTR1_search(Name, p1s, p1e, p2s, p2e, r1, r2, r3, r4, r5, r6):
 			#vntr_count = difflib.get_close_matches(r2, fragment, cutoff = 0.85)
 			vntr_count = fragment.count(r2) + fragment.count(r4) + fragment.count(r6)
 			MLVA_code.append(vntr_count)
+		print >>seq, ">%s" % (name+"_" +head), "\n" + fragment
+		print >>q, "Fragment locus:", fp1,"-",fp2
+		print >>q, "Fragment size:", fDistance	
+		print >>q, "Number of repeats:", vntr_count
 	else:
 		pass
 
@@ -214,9 +204,6 @@ def VNTR1_search(Name, p1s, p1e, p2s, p2e, r1, r2, r3, r4, r5, r6):
 		pattern = re.compile("|".join(replace.keys()))
 		rev_text = pattern.sub(lambda m: replace[re.escape(m.group(0))], fragment)
 		fragment_rev = rev_text[::-1]
-		print >>seq, ">%s" % (Name), "\n" + fragment_rev
-		print >>q, "VNTR Fragment locus:", rp1, "-", rp2
-		print >>q, "VNTR fragment size:", rDistance	
 		if r1 or r3 or r5 in fragment_rev:	
 			vntr_count = fragment_rev.count(r1) + \
 			fragment_rev.count(r3) + fragment_rev.count(r5)
@@ -225,24 +212,28 @@ def VNTR1_search(Name, p1s, p1e, p2s, p2e, r1, r2, r3, r4, r5, r6):
 			vntr_count = fragment_rev.count(r2) + \
 			fragment_rev.count(r4) + fragment_rev.count(r6)
 			MLVA_code.append(vntr_count)
+		print >>seq, ">%s" % (name+"_" +head), "\n" + fragment_rev
+		print >>q, "Fragment locus:", rp1, "-", rp2
+		print >>q, "Fragment size:", rDistance	
+		print >>q, "Number of repeats:", vntr_count	
 	else:
 		pass
 
-def VNTR_search(Name, p1s, p1e, p2s, p2e, r1, r2):
-	print>>q, "\n >>>" + Name + "\n"
+def VNTR_search(head, p1s, p1e, p2s, p2e, r1, r2):
+	print>>q, "\n>>> " + head
 	if p1s and p1e in all_dna:
 		for x in re.finditer(p1s, all_dna):
 			fp1 = int(x.start())
 		for y in re.finditer(p1e, all_dna):
 			fp2 = int(y.end())
 		fDistance = fp2 - fp1
-		fragment = all_dna[fp1:fp2]
-		print >>seq, ">%s" % (Name), "\n" + fragment
-		print >>q, "VNTR Fragment locus:", fp1, "-", fp2
-		print >>q, "VNTR fragment size:", fDistance	
+		fragment = all_dna[fp1:fp2]	
 		vntr_count = fragment.count(r1) or fragment_rev.count(r2)
 		MLVA_code.append(vntr_count)
-
+		print >>seq, ">%s" % (name+"_" +head), "\n" + fragment
+		print >>q, "Fragment locus:", fp1, "-", fp2
+		print >>q, "Fragment size:", fDistance
+		print >>q, "Number of repeats:", vntr_count
 	else:
 		pass
 
@@ -258,11 +249,12 @@ def VNTR_search(Name, p1s, p1e, p2s, p2e, r1, r2):
 		pattern = re.compile("|".join(replace.keys()))
 		rev_text = pattern.sub(lambda m: replace[re.escape(m.group(0))], fragment)
 		fragment_rev = rev_text[::-1]
-		print >>seq, ">%s" % (Name), "\n" + fragment_rev
-		print >>q, "VNTR Fragment locus:", rp1, "-", rp2
-		print >>q, "VNTR fragment size:", rDistance
 		vntr_count = fragment_rev.count(r1) or fragment_rev.count(r2)
 		MLVA_code.append(vntr_count)
+		print >>seq, ">%s" % (name+"_" +head), "\n" + fragment_rev
+		print >>q, "Fragment locus:", rp1, "-", rp2
+		print >>q, "Fragment size:", rDistance
+		print >>q, "Number of repeats:", vntr_count
 	else:
 		pass
 
@@ -289,14 +281,6 @@ def VNTR3_search(p1s, p1e, p2s, p2e, r1, r2):
 			fDistance2 = V3_end_2 - V3_start_2
 			fragment1 = all_dna[V3_start_1:V3_end_1]
 			fragment2 = all_dna[V3_start_2:V3_end_2]
-			print >>seq, ">%s" % ("VNTR3a"), "\n" + fragment1
-			print >>seq, ">%s" % ("VNTR3b"), "\n" + fragment2
-			print >>q, "\n >>> VNTR3a" + "\n"
-			print >>q, "VNTR3a Fragment locus:", V3_start_1, "-", V3_end_1
-			print >>q, "VNTR3a fragment size:", fDistance1
-			print >>q, "\n >>> VNTR3_2" + "\n"
-			print >>q, "VNTR3b Fragment locus:", V3_start_2, "-", V3_end_2
-			print >>q, "VNTR3b fragment size:", fDistance2
 			vntr_count1 = fragment1.count(r1) or fragment1.count(r2)
 			vntr_count2 = fragment2.count(r1) or fragment2.count(r2)
 			if vntr_count1 == vntr_count2 and fix == False:
@@ -308,23 +292,34 @@ def VNTR3_search(p1s, p1e, p2s, p2e, r1, r2):
 			else:
 				MLVA_code.append(vntr_count2)
 				MLVA_code.append(vntr_count1)
+			print >>seq, ">%s" % (name+"_VNTR3a"), "\n" + fragment1
+			print >>seq, ">%s" % (name+"_VNTR3b"), "\n" + fragment2
+			print >>q, "\n>>> VNTR3a"
+			print >>q, "Fragment locus:", V3_start_1, "-", V3_end_1
+			print >>q, "Fragment size:", fDistance1
+			print >>q, "\n>>> VNTR3b"
+			print >>q, "Fragment locus:", V3_start_2, "-", V3_end_2
+			print >>q, "Fragment size:", fDistance2
+			print >>q, "Number of repeats:", vntr_count
 
 		elif (len(V3p1_start) == 1) and (p2s not in all_dna):
 			V3_start_single_1 = int(V3p1_start[0])
 			V3_end_single_1 = int(V3p1_end[0])
 			fDistance = V3_end_single_1 - V3_start_single_1
 			fragment = all_dna[V3_start_single_1:V3_end_single_1]
-			print >>seq, ">%s" % ("VNTR3_1"), "\n" + fragment
-			print >>q, "\n >>> VNTR3_1" + "\n"
-			print >>q, "VNTR3a Fragment locus:", V3_start_single_1, "-",\
-			V3_end_single_1
-			print >>q, "VNTR3a fragment size:", fDistance
-			print >>q, "\n >>> VNTR3_2" + "\n"
-			print >>q, "VNTR3b Fragment locus: 0"
-			print >>q, "VNTR3b fragment size: 0"
 			vntr_count1 = fragment.count(r1) or fragment.count(r2)
 			MLVA_code.append(vntr_count1)
 			MLVA_code.append(0)
+			print >>seq, ">%s" % (name+"_VNTR3a"), "\n" + fragment
+			print >>q, "\n>>> VNTR3a"
+			print >>q, "Fragment locus:", V3_start_single_1, "-",\
+			V3_end_single_1
+			print >>q, "Fragment size:", fDistance
+			print >>q, "Number of repeats:", vntr_count
+			print >>q, "\n>>> VNTR3b"
+			print >>q, "Fragment locus: 0"
+			print >>q, "Fragment size: 0"
+			print >>q, "Number of repeats: 0"
 		else:
 			pass
 
@@ -353,15 +348,7 @@ def VNTR3_search(p1s, p1e, p2s, p2e, r1, r2):
 			rev_text2 = pattern.sub(lambda m: replace[re.escape(m.group(0))], fragment2)
 			fragment1rev = rev_text1[::-1]
 			fragment2rev = rev_text2[::-1]
-
-			print >>seq, ">%s" % ("VNTR3a"), "\n" + fragment1rev
-			print >>seq, ">%s" % ("VNTR3b"), "\n" + fragment2rev
-			print >>q, "\n >>> VNTR3a" + "\n"
-			print >>q, "VNTR3a Fragment locus:", V3_start_1, "-", V3_end_1
-			print >>q, "VNTR3a fragment size:", fDistance1
-			print >>q, "\n >>> VNTR3_2" + "\n"
-			print >>q, "VNTR3b Fragment locus:", V3_start_2, "-", V3_end_2
-			print >>q, "VNTR3b fragment size:", fDistance2
+			
 			vntr_count1 = fragment1rev.count(r1) or fragment1rev.count(r2)
 			vntr_count2 = fragment2rev.count(r2) or fragment2rev.count(r2)
 			if vntr_count1 == vntr_count2 and fix == False:
@@ -373,6 +360,16 @@ def VNTR3_search(p1s, p1e, p2s, p2e, r1, r2):
 			else:
 				MLVA_code.append(vntr_count2)
 				MLVA_code.append(vntr_count1)
+			print >>seq, ">%s" % (name+"_VNTR3a"), "\n" + fragment1rev
+			print >>seq, ">%s" % (name+"_VNTR3b"), "\n" + fragment2rev
+			print >>q, "\n>>> VNTR3a"
+			print >>q, "Fragment locus:", V3_start_1, "-", V3_end_1
+			print >>q, "Fragment size:", fDistance1
+			print >>q, "Number of repeats:", vntr_count
+			print >>q, "\n>>> VNTR3b"
+			print >>q, "Fragment locus:", V3_start_2, "-", V3_end_2
+			print >>q, "Fragment size:", fDistance2	
+			print >>q, "Number of repeats:", vntr_count
 		elif (len(V3p2_start) == 1) and (p1s not in all_dna):
 			V3_start_single_2 = int(V3p2_start[0])
 			V3_end_single_2 = int(V3p2_end[0])
@@ -383,18 +380,20 @@ def VNTR3_search(p1s, p1e, p2s, p2e, r1, r2):
 			pattern = re.compile("|".join(replace.keys()))
 			rev_text1 = pattern.sub(lambda m: replace[re.escape(m.group(0))], fragment)
 			fragment_rev = rev_text1[::-1]
-
-			print >>seq, ">%s" % ("VNTR3a"), "\n" + fragment_rev
-			print >>q, "\n >>> VNTR3a" + "\n"
-			print >>q, "VNTR3a Fragment locus:", V3_start_single_2, "-",\
-			V3_end_single_2
-			print >>q, "VNTR3a fragment size:", rDistance
-			print >>q, "\n >>> VNTR3b" + "\n"
-			print >>q, "VNTR3b Fragment locus: 0"
-			print >>q, "VNTR3b fragment size: 0"
 			vntr_count2 = fragment_rev.count(r1) or fragment_rev.count(r2)
 			MLVA_code.append(vntr_count2)
 			MLVA_code.append(0)
+			print >>seq, ">%s" % (name+"_VNTR3a"), "\n" + fragment_rev
+			print >>q, "\n>>> VNTR3a"
+			print >>q, "Fragment locus:", V3_start_single_2, "-",\
+			V3_end_single_2
+			print >>q, "Fragment size:", rDistance
+			print >>q, "Number of repeats:", vntr_count
+			print >>q, "\n>>> VNTR3b"
+			print >>q, "Fragment locus: 0"
+			print >>q, "Fragment size: 0"
+			print >>q, "Number of repeats: 0"
+			
 		else:
 			pass
 			
@@ -426,13 +425,6 @@ def VNTR3_search(p1s, p1e, p2s, p2e, r1, r2):
 		rev_text2 = pattern.sub(lambda m: replace[re.escape(m.group(0))], fragment2)
 		fragment2rev = rev_text2[::-1]
 
-		print >>seq, ">%s" % ("VNTR3a"), "\n" + fragment2rev
-		print >>q, "\n >>> VNTR3a" + "\n"
-		print >>q, "VNTR3a Fragment locus:", V3_start_1, "-", V3_end_1
-		print >>q, "VNTR3a fragment size:", fDistance
-		print >>q, "\n >>> VNTR3b" + "\n"
-		print >>q, "VNTR3b Fragment locus:", V3_start_2, "-", V3_end_2
-		print >>q, "VNTR3b fragment size:", rDistance
 		vntr_count1 = fragment1.count(r1) or fragment1.count(r2)
 		vntr_count2 = fragment2.count(r1) or fragment2.count(r2)
 		if vntr_count1 == vntr_count2 and fix == False:
@@ -444,42 +436,57 @@ def VNTR3_search(p1s, p1e, p2s, p2e, r1, r2):
 		else:
 			MLVA_code.append(vntr_count2)
 			MLVA_code.append(vntr_count1)
+		print >>seq, ">%s" % (name+"_VNTR3a"), "\n" + fragment2rev
+		print >>q, "\n>>> VNTR3a"
+		print >>q, "Fragment locus:", V3_start_1, "-", V3_end_1
+		print >>q, "Fragment size:", fDistance
+		print >>q, "Number of repeats:", vntr_count1
+		print >>q, "\n>>> VNTR3b"
+		print >>q, "Fragment locus:", V3_start_2, "-", V3_end_2
+		print >>q, "Fragment size:", rDistance
+		print >>q, "Number of repeats:", vntr_count2
 	else:
 		pass		
 
 #------------------------------------------------------------------------------
 
-VNTR1_search(Name = "VNTR1", p1s = V1p1s,p1e = V1p1e, p2s = V1p2s, \
+VNTR1_search(head = "VNTR1", p1s = V1p1s,p1e = V1p1e, p2s = V1p2s, \
 			p2e = V1p2e, r1 = VNTR1wt1,r2 = VNTR1wt2, r3 = VNTR1mut1_13, \
 			r4 = VNTR1mut2_13, r5 = VNTR1mut1_1315,r6 = VNTR1mut2_1315)
 
-VNTR3_search(p1s = V3p1s, p1e = V3p1e, p2s = V3p2s, p2e = V3p2e, r1 = VNTR3_1, r2 = VNTR3_2)
+VNTR3_search(p1s = V3p1s, p1e = V3p1e, p2s = V3p2s, p2e = V3p2e, \
+			r1 = VNTR3_1, r2 = VNTR3_2)
 
-VNTR_search(Name = "VNTR4", p1s = V4p1s, p1e = V4p1e, p2s = V4p2s, p2e = V4p2e, r1 = VNTR4_1, r2 = VNTR4_2)
+VNTR_search(head = "VNTR4", p1s = V4p1s, p1e = V4p1e, p2s = V4p2s, \
+			p2e = V4p2e, r1 = VNTR4_1, r2 = VNTR4_2)
 
-VNTR_search(Name = "VNTR5", p1s = V5p1s, p1e = V5p1e, p2s = V5p2s, p2e = V5p2e, r1 = VNTR5_1, r2 = VNTR5_2)
+VNTR_search(head = "VNTR5", p1s = V5p1s, p1e = V5p1e, p2s = V5p2s, \
+			p2e = V5p2e, r1 = VNTR5_1, r2 = VNTR5_2)
 
-VNTR_search(Name = "VNTR6", p1s = V6p1s, p1e = V6p1e, p2s = V6p2s, p2e = V6p2e, r1 = VNTR6_1, r2 = VNTR6_2)
+VNTR_search(head = "VNTR6", p1s = V6p1s, p1e = V6p1e, p2s = V6p2s, \
+			p2e = V6p2e, r1 = VNTR6_1, r2 = VNTR6_2)
 
 #------------------------------------------------------------------------------
-#Output information. Modifies and sends results to output files.
-print >>q, "\n>>> MLVA Code: ", MLVA_code
 
 #Send MLVA number to output file
-
 searchfile = open("%s/MLVAdatabase.txt" % mlvaDir, 'r')
-'''
-s = open("%s/MLVA/summary.txt" % outDir, "a")
-print >>s,name, ",", str(MLVA_code)
-'''
-for line in searchfile:
-	if str(MLVA_code) in line:
-		s = open("%s/MLVA/summary.txt" % outDir, "a")
-		print >>s, name,',', line.rstrip('\n')
-	else:
-		pass
 
-#print >>s, name, MLVA_code
-print ">>> Removing temperary files..."
-os.remove("%s/MLVA/MLVA_data/%s" % (outDir, outFile))
+for line in searchfile:
+	s = open("%s/MLVA/summary.txt" % outDir, "a")
+	if str(MLVA_code) in line:
+		print >>s, name,':', line.rstrip('\n')
+		print >>q, "\n>>> MLVA Type: ", line.rstrip('\n')
+		test = True
+	else:
+		test = False
+
+if test == False:
+	print >>s, name,':', "MLVA Type not Found"
+	print >>q, "\n>>> MLVA Type: MLVA Type not Found"
+
+if keep == False:
+	print ">>> Removing temperary files..."
+	os.remove("%s/MLVA/MLVA_data/%s" % (outDir, outFile))
+else:
+	pass
 print ">>> Search Complete."
